Edit and add components for interpreter minus OOP.
Cleaner is a work-in-progress GC-like object to deallocate the memory originally allocated for the Stmt nodes in the Parser.
Very late commit due to multiple changes, additions, and corrections being needed in several parts simultaneously.
Interpreter currently supports all the basic Lox features (plus extras, like lambdas and some helpful warnings) sans classes (the next thing to implement).
There are also some issues carried over from the jlox implementation which should be fixed once the basic features have been implemented and debugged thoroughly.

+++ src/LoxFunction.cpp
@@ -29,7 +29,7 @@ Object LoxFunction::call(Interpreter interpreter, std::vector<Object> arguments)
     if (declaration.params != nullptr)
     {
         vT params = *declaration.params;
-        for (size_t i = 0; i < declaration.params->size(); i++)
+        for (int i = 0; i < (int) declaration.params->size(); i++)
             environment.define(params[i].lexeme, arguments[i]);
     }
 
@@ -59,7 +59,7 @@ bool LoxFunction::isGetter()
 int LoxFunction::arity()
 {
     if (declaration.params == nullptr) return 0;
-    return declaration.params->size();
+    return (int) declaration.params->size();
 }
 
 std::string LoxFunction::toString()
